apiproperty and renaming
