+++ apps/server/src/modules/system/controller/dto/public-system-response.ts
